+++ circuit-retained/src/commonMain/kotlin/com/slack/circuit/retained/Continuity.kt
@@ -8,11 +8,9 @@ import androidx.compose.runtime.Composable
 @Composable
 public expect fun continuityRetainedStateRegistry(
   key: String = Continuity.KEY,
-  canRetainChecker: CanRetainChecker = rememberContinuityCanRetainChecker(),
+  canRetainChecker: CanRetainChecker = CanRetainChecker.Always,
 ): RetainedStateRegistry
 
 public object Continuity {
   public const val KEY: String = "CircuitContinuity"
 }
-
-@Composable public expect fun rememberContinuityCanRetainChecker(): CanRetainChecker