+++ circuit-retained/src/jvmMain/kotlin/com/slack/circuit/retained/Continuity.jvm.kt
@@ -11,9 +11,3 @@ public actual fun continuityRetainedStateRegistry(
 ): RetainedStateRegistry {
   return rememberRetainedStateRegistry(key = key, canRetainChecker = canRetainChecker)
 }
-
-/** Checks whether or not we can retain in the current composable context. */
-@Composable
-public actual fun rememberContinuityCanRetainChecker(): CanRetainChecker {
-  return CanRetainChecker.Always
-}